postgresql uses NUMERIC

+++ src/sql/postgresql/sessions_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE sessions (
   targetcompid VARCHAR(64) NOT NULL,
   session_qualifier VARCHAR(64) NOT NULL,
   creation_time TIMESTAMP NOT NULL,
-  incoming_seqnum INTEGER NOT NULL, 
-  outgoing_seqnum INTEGER NOT NULL,
+  incoming_seqnum NUMERIC NOT NULL, 
+  outgoing_seqnum NUMERIC NOT NULL,
   PRIMARY KEY (beginstring, sendercompid, targetcompid, session_qualifier)
 );
